requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
itsdangerous>=2.0
Jinja2>=3.0
click>=8.0
Flask-Dance==5.0.0
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.46
google-cloud-firestore==2.7.0
pytz==2023.3
Flask-Session
python-dateutil
python-dotenv